refactor: Move luis exceptions to Recognizer namespace
