+++ src/Common/Recognizer/InvalidWorkedPeriodInstanceException.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 
-namespace Bot.Clockify
+namespace Bot.Common.Recognizer
 {
-    // TODO Rearrange exceptions so that Fill and reports do not share this one
     public class InvalidWorkedPeriodInstanceException : Exception
     {
         public InvalidWorkedPeriodInstanceException(string message) : base(message)
